+++ R/PipeOpBLSmote.R
@@ -46,6 +46,10 @@
 #' * `quiet` :: `logical(1)` \cr
 #'   Whether to suppress printing status during training. Initialized to `TRUE`.
 #'
+#' @section Internals:
+#' If a target level is unobserved during training, no synthetic data points will be generated for that class.
+#' No error is raised; the unobserved class is simply ignored.
+#'
 #' @section Fields:
 #' Only fields inherited from [`PipeOp`].
 #'
@@ -110,14 +114,18 @@ PipeOpBLSmote = R6Class("PipeOpBLSmote",
 
       # Calculate synthetic data
       dt = task$data(cols = cols)
+      # Remove unseen target levels, see #881
+      # Don't need to re-add them later since we don't touch task here
+      target = droplevels(task$truth())
+
       if (self$param_set$get_values()$quiet) {
         utils::capture.output({
-          st = setDT(invoke(smotefamily::BLSMOTE, X = dt, target = task$truth(),
+          st = setDT(invoke(smotefamily::BLSMOTE, X = dt, target = target,
                             .args = self$param_set$get_values(tags = "blsmote"),
                             .opts = list(warnPartialMatchArgs = FALSE))$syn_data)  # BLSMOTE uses partial arg matching internally
         })
       } else {
-        st = setDT(invoke(smotefamily::BLSMOTE, X = dt, target = task$truth(),
+        st = setDT(invoke(smotefamily::BLSMOTE, X = dt, target = target,
                           .args = self$param_set$get_values(tags = "blsmote"),
                           .opts = list(warnPartialMatchArgs = FALSE))$syn_data)
       }