+++ inst/testthat/helper_functions.R
@@ -210,6 +210,7 @@ expect_pipeop_class = function(poclass, constargs = list(), check_ps_default_val
 #  - predicting on task that has different column layout than training gives an error
 #  - training on task with no columns returns task with no columns
 #  - training on task when affect_columns is FALSE does not change task
+#  - training on classification tasks with empty target level
 #  - predicting on task with no columns returns task with no columns (if training happened on the same)
 #  - predicting on task with no rows returns task with no rows (if predict_rows_independent)
 #  - predicting without target column works
@@ -230,7 +231,6 @@ expect_datapreproc_pipeop_class = function(poclass, constargs = list(), task,
   # NOTE
   # The 'tolerance' parameter is not used in many places yet; if tolerance becomes a problem, add the
   # 'tolerance = tolerance' argument to `expect_equal`.
-
 
   original_clone = task$clone(deep = TRUE)
   expect_shallow_clone(task, original_clone)
@@ -401,12 +401,22 @@ expect_datapreproc_pipeop_class = function(poclass, constargs = list(), task,
 
   }
 
+  # test that training works for classification tasks with empty target level, #881
+  if (po$input$train != "Task" && inherits(task, "TaskClassif")) {
+    task3 = task$clone(deep = TRUE)
+    new_levels = list(c(task3$levels(cols = task3$target_names)[[task3$target_names]], "empty_level"))
+    task3$set_levels(set_names(new_levels, task3$target_names))
+    expect_no_error({
+      train_out = po$train(list(task3))
+      expect_contains(train_out[[1L]]$levels(cols = task$target_names)[[1L]], "empty_level")
+    })
+  }
+
   po$train(list(task))
 
   norowtask = task$clone(deep = TRUE)$filter(integer(0))
   whichrow = task$row_ids[[sample.int(task$nrow, 1)]]
   onerowtask = task$clone(deep = TRUE)$filter(whichrow)
-
 
   predicted = po$predict(list(norowtask))[[1]]
   if (predict_rows_independent) {
@@ -458,7 +468,7 @@ expect_datapreproc_pipeop_class = function(poclass, constargs = list(), task,
   taskouttrain = po$train(list(tasktrain))[[1L]]
   taskoutpredict = po$predict(list(taskpredict))[[1L]]
 
-  # other columns like weights are present during traing but not during predict
+  # other columns like weights are present during training but not during predict
   cols = unname(unlist(taskouttrain$col_roles[c("feature", "target")]))
   dtrain = taskouttrain$internal_valid_task$data(cols = cols)
   dpredict = taskoutpredict$data(cols = cols)